requirements.txt
pbr
openstacksdk
keystoneauth1
keystonemiddleware
oslo.context
oslo.config
oslo.log
oslo.messaging
oslo.policy
flask>=1.0.2
marshmallow>=3.4.0
flask_migrate>=2.1.1
flask_restful>=0.3.6
flask_sqlalchemy>=2.4.1
flask_marshmallow>=0.11.0
marshmallow_sqlalchemy>=0.22.3
cotyledon
pymysql
python-memcached
prometheus_client
